refactor: move highlight to color.ts
